Move currencies request to the dahsboard container - Move currency request loader to the calc section
