+++ src/App.tsx
@@ -1,40 +1,18 @@
-import React, { useEffect, useState } from 'react'
-import './App.scss'
-import { ApiRate, Dashboard } from './components/Dashboard'
-import MainHeader from './components/MainHeaderComponent/MainHeaderComponent'
-import { Layout, Spin } from 'antd'
+import React from 'react'
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
+import { Layout } from 'antd'
 
 import About from './components/AboutComponent/AboutComponent'
 import Contact from './components/ContactComponent/ContactComponent'
 import UndefinedRoute from './components/UndefinedRouteComponent/UndefinedRouteComponent'
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import { getCurrencyValues } from './services/CurrencyService'
-import { from } from 'rxjs'
+import MainHeader from './components/MainHeaderComponent/MainHeaderComponent'
+import DashboardContainer from "./components/Dashboard/Dashboard.container"
+
+import './App.scss'
 
 const {Header, Content} = Layout
 
-const resource = getCurrencyValues()
-const $resourceObservable = from(resource)
-
-type Response = {
-    rates: ApiRate[];
-    date: string
-}
-
 const App: React.FC = () => {
-
-    const [apiRates, setApiRates] = useState<ApiRate[]>([])
-    const [date, setCurrentDate] = useState('')
-    const [dataReady, setDataReady] = useState(false)
-
-    useEffect(() => {
-        $resourceObservable.subscribe((response: Response) => {
-            setApiRates(response.rates)
-            setCurrentDate(response.date)
-            setDataReady(true)
-        })
-    }, [])
-
     return (
         <div className="cc-theme-purple">
             <Router basename={process.env.PUBLIC_URL}>
@@ -49,23 +27,13 @@ const App: React.FC = () => {
                             path="/"
                         >
                             <Content className="main-layout__content">
-                                {!dataReady
-                                    ? <Spin/>
-                                    : <Dashboard
-                                        rates={apiRates}
-                                        date={date}
-                                    />
-                                }
+                                <DashboardContainer/>
                             </Content>
                         </Route>
 
                         <Route path="/about">
                             <Content className="main-layout__content">
-                                {!dataReady
-                                    ? <Spin/>
-                                    :
-                                    <About/>
-                                }
+                                <About/>
                             </Content>
                         </Route>
