+++ src/components/Dashboard/Dashboard.scss
@@ -1,6 +1,6 @@
-@import "../styles/utils/variables";
-@import "../styles/utils/fonts";
-@import "../styles/utils/mixins";
+@import "../../styles/utils/variables";
+@import "../../styles/utils/fonts";
+@import "../../styles/utils/mixins";
 
 .converter {
   display: grid;
@@ -18,7 +18,7 @@
     padding: $gap-big 0;
     background-color: $cc-purple;
     background-image: linear-gradient(45deg, transparentize($cc-purple-dark, .07) 10%, transparentize($cc-purple, .07)),
-    url("../assets/img/bg-calc-info.jpg");
+    url("../../assets/img/bg-calc-info.jpg");
     background-position: center;
     background-size: cover;
     color: $cc-white;
@@ -60,6 +60,7 @@
     z-index: 2;
     grid-column: 6 / 13;
     grid-row: 2;
+    min-height: 470px;
     padding: 40px;
     background-color: $cc-white;
     text-align: center;
