+++ src/components/Dashboard/Dashboard.tsx
@@ -1,20 +1,22 @@
-import { Button, Input } from 'antd'
+import React, { useEffect, useState } from 'react'
+import {Button, Input, Spin} from 'antd'
+import { SwapOutlined } from "@ant-design/icons/lib";
+import { Helmet } from 'react-helmet'
+import { Trans, useTranslation } from "react-i18next";
+
+import { initialSelectFromValue, initialSelectToValue, inputPlaceholder } from '../../constants/Variables'
+import {getParsedNumber} from "../../utils/number";
+import {CurrencySelect} from "../CurrecySelectComponent/SelectComponent";
+import {RateDisplay} from "../DateAndRateDisplayComponent/DateAndRateDisplayComponent";
 
 import './Dashboard.scss'
-import React, { useEffect, useState } from 'react'
-import { CurrencySelect } from './CurrecySelectComponent/SelectComponent'
-import { RateDisplay } from './DateAndRateDisplayComponent/DateAndRateDisplayComponent'
-import { initialSelectFromValue, initialSelectToValue, inputPlaceholder } from '../constants/Variables'
-import { Helmet } from 'react-helmet'
-import { getParsedNumber } from "../utils/number"
-import { SwapOutlined } from "@ant-design/icons/lib";
-import { Trans, useTranslation } from "react-i18next";
 
 const userInputRegex = new RegExp('^\\d+([,.]\\d{0,2})?$')
 
 type Props = {
     rates: ApiRate[]
     date: string
+    dataReady: boolean
 }
 
 export type ApiRate = {
@@ -25,8 +27,7 @@ export type ApiRate = {
 
 type SelectedValue = ApiRate | undefined
 
-export const Dashboard: React.FC<Props> = ({rates, date}) => {
-
+export const Dashboard: React.FC<Props> = ({rates, date, dataReady}) => {
     const {t} = useTranslation();
 
     const [fromCurrency, setFromCurrency] = useState('')
@@ -102,52 +103,58 @@ export const Dashboard: React.FC<Props> = ({rates, date}) => {
                 <h2 className="converter__calc__header">
                     {t('ConvertCurrencyCalculatorHeader')}
                 </h2>
+                {!dataReady ?
+                  <Spin/>
+                  : <>
+                      <div className="converter__calc__group">
+                          <CurrencySelect
+                            value={fromCurrency}
+                            name="fromCurrency"
+                            currencyRates={rates}
+                            onChange={setFromCurrency}
+                          />
+                          <label htmlFor="inputValue" className="sr-only">Input Value</label>
+                          <Input
+                            placeholder={inputPlaceholder}
+                            value={userValue}
+                            id="inputValue"
+                            disabled={!dataReady}
+                            onChange={onChangeValue}
+                          />
+                      </div>
 
-                <div className="converter__calc__group">
-                    <CurrencySelect
-                        value={fromCurrency}
-                        name="fromCurrency"
-                        onChange={setFromCurrency}
-                        currencyRates={rates}
-                    />
-                    <label htmlFor="inputValue" className="sr-only">Input Value</label>
-                    <Input
-                        placeholder={inputPlaceholder}
-                        value={userValue}
-                        id="inputValue"
-                        onChange={onChangeValue}
-                    />
-                </div>
+                      <Button
+                        type="primary"
+                        className="btn-swap cc-btn--gradient"
+                        disabled={!fromCurrency || !toCurrency}
+                        onClick={onCurrencySwap}
+                      >
+                          <SwapOutlined
+                            rotate={90}
+                            className="btn-swap-icon"
+                          />
+                      </Button>
 
-                <Button
-                    type="primary"
-                    className="btn-swap cc-btn--gradient"
-                    disabled={!fromCurrency || !toCurrency}
-                    onClick={onCurrencySwap}
-                >
-                    <SwapOutlined
-                        rotate={90}
-                        className="btn-swap-icon"
-                    />
-                </Button>
+                      <div className="converter__calc__group">
+                          <CurrencySelect
+                            value={toCurrency}
+                            name="toCurrency"
+                            currencyRates={rates}
+                            onChange={setToCurrency}
+                          />
+                          <label htmlFor="outputValue" className="sr-only">Output Value</label>
+                          <Input
+                            placeholder={inputPlaceholder}
+                            id="outputValue"
+                            value={converterValue}
+                          />
+                      </div>
 
-                <div className="converter__calc__group">
-                    <CurrencySelect
-                        value={toCurrency}
-                        name="toCurrency"
-                        onChange={setToCurrency}
-                        currencyRates={rates}
-                    />
-                    <label htmlFor="outputValue" className="sr-only">Output Value</label>
-                    <Input
-                      placeholder={inputPlaceholder}
-                      id="outputValue"
-                      value={converterValue}
-                    />
-                </div>
-                <div className="converter__calc__rate">
-                    <p>{RateDisplay(prepareExchangeRateInformation())}</p>
-                </div>
+                      <div className="converter__calc__rate">
+                          <p>{RateDisplay(prepareExchangeRateInformation())}</p>
+                      </div>
+                  </>
+                }
             </div>
         </div>
     )
